Introduce a new `src/output` module
