+++ .github/workflows/website-deploy.yml
@@ -29,6 +29,7 @@ jobs:
           -DCMAKE_BUILD_TYPE:STRING=Release
           -DBLAZE_COMPILER:BOOL=OFF
           -DBLAZE_EVALUATOR:BOOL=OFF
+          -DBLAZE_OUTPUT:BOOL=OFF
           -DBLAZE_LINTER:BOOL=OFF
           -DBLAZE_TESTS:BOOL=OFF
           -DBLAZE_DOCS:BOOL=ON
