+++ CMakeLists.txt
@@ -7,6 +7,7 @@ list(APPEND CMAKE_MODULE_PATH "${PROJECT_SOURCE_DIR}/cmake")
 # Options
 option(BLAZE_COMPILER "Build the Blaze compiler library" ON)
 option(BLAZE_EVALUATOR "Build the Blaze evaluator library" ON)
+option(BLAZE_OUTPUT "Build the Blaze output formats library" ON)
 option(BLAZE_LINTER "Build the Blaze linter rule library" ON)
 option(BLAZE_TESTS "Build the Blaze tests" OFF)
 option(BLAZE_BENCHMARK "Build the Blaze benchmarks" OFF)
@@ -48,6 +49,10 @@ if(BLAZE_EVALUATOR)
   add_subdirectory(src/evaluator)
 endif()
 
+if(BLAZE_OUTPUT)
+  add_subdirectory(src/output)
+endif()
+
 if(BLAZE_LINTER)
   add_subdirectory(src/linter)
 endif()
@@ -85,6 +90,10 @@ if(BLAZE_TESTS)
 
   if(BLAZE_EVALUATOR)
     add_subdirectory(test/evaluator)
+  endif()
+
+  if(BLAZE_OUTPUT)
+    add_subdirectory(test/output)
   endif()
 
   if(BLAZE_LINTER)
