+++ config.cmake.in
@@ -6,6 +6,7 @@ list(APPEND BLAZE_COMPONENTS ${blaze_FIND_COMPONENTS})
 if(NOT BLAZE_COMPONENTS)
   list(APPEND BLAZE_COMPONENTS compiler)
   list(APPEND BLAZE_COMPONENTS evaluator)
+  list(APPEND BLAZE_COMPONENTS output)
   list(APPEND BLAZE_COMPONENTS linter)
 endif()
 
@@ -21,7 +22,11 @@ foreach(component ${BLAZE_COMPONENTS})
   elseif(component STREQUAL "linter")
     include("${CMAKE_CURRENT_LIST_DIR}/sourcemeta_blaze_evaluator.cmake")
     include("${CMAKE_CURRENT_LIST_DIR}/sourcemeta_blaze_compiler.cmake")
+    include("${CMAKE_CURRENT_LIST_DIR}/sourcemeta_blaze_output.cmake")
     include("${CMAKE_CURRENT_LIST_DIR}/sourcemeta_blaze_linter.cmake")
+  elseif(component STREQUAL "output")
+    include("${CMAKE_CURRENT_LIST_DIR}/sourcemeta_blaze_evaluator.cmake")
+    include("${CMAKE_CURRENT_LIST_DIR}/sourcemeta_blaze_output.cmake")
   else()
     message(FATAL_ERROR "Unknown Blaze component: ${component}")
   endif()
