+++ src/output/include/sourcemeta/blaze/output.h
@@ -1,0 +1,8 @@
+#ifndef SOURCEMETA_BLAZE_OUTPUT_H_
+#define SOURCEMETA_BLAZE_OUTPUT_H_
+
+#include <sourcemeta/blaze/output_simple.h>
+#include <sourcemeta/blaze/output_standard.h>
+#include <sourcemeta/blaze/output_trace.h>
+
+#endif
