+++ test/compiler/CMakeLists.txt
@@ -1,10 +1,6 @@
 sourcemeta_googletest(NAMESPACE sourcemeta PROJECT blaze NAME compiler
   FOLDER "Blaze/Compiler"
   SOURCES
-    compiler_output_simple_test.cc
-    compiler_output_trace_test.cc
-    compiler_output_standard_flag_test.cc
-    compiler_output_standard_basic_test.cc
     compiler_unevaluated_2019_09_test.cc
     compiler_unevaluated_2020_12_test.cc
     compiler_json_test.cc
