+++ test/output/output_simple_test.cc
@@ -2,6 +2,7 @@
 
 #include <sourcemeta/blaze/compiler.h>
 #include <sourcemeta/blaze/evaluator.h>
+#include <sourcemeta/blaze/output.h>
 
 #include <sourcemeta/core/jsonpointer.h>
 #include <sourcemeta/core/jsonschema.h>
@@ -54,7 +55,7 @@
               (expected_value));                                               \
   }
 
-TEST(Compiler_output_simple, success_string_1) {
+TEST(Output_simple, success_string_1) {
   const sourcemeta::core::JSON schema{sourcemeta::core::parse_json(R"JSON({
     "$schema": "https://json-schema.org/draft/2020-12/schema",
     "type": "string"
@@ -80,7 +81,7 @@ TEST(Compiler_output_simple, success_string_1) {
   EXPECT_ANNOTATION_COUNT(output, 0);
 }
 
-TEST(Compiler_output_simple, fail_meaningless_if_1) {
+TEST(Output_simple, fail_meaningless_if_1) {
   const sourcemeta::core::JSON schema{sourcemeta::core::parse_json(R"JSON({
     "$schema": "https://json-schema.org/draft/2019-09/schema",
     "properties": {
@@ -129,7 +130,7 @@ TEST(Compiler_output_simple, fail_meaningless_if_1) {
   EXPECT_ANNOTATION_COUNT(output, 0);
 }
 
-TEST(Compiler_output_simple, success_dynamic_anchor_1) {
+TEST(Output_simple, success_dynamic_anchor_1) {
   const sourcemeta::core::JSON schema{sourcemeta::core::parse_json(R"JSON({
     "$schema": "https://json-schema.org/draft/2020-12/schema",
     "$dynamicRef": "#foo",
@@ -161,7 +162,7 @@ TEST(Compiler_output_simple, success_dynamic_anchor_1) {
   EXPECT_ANNOTATION_COUNT(output, 0);
 }
 
-TEST(Compiler_output_simple, success_oneof_1) {
+TEST(Output_simple, success_oneof_1) {
   const sourcemeta::core::JSON schema{sourcemeta::core::parse_json(R"JSON({
     "$schema": "https://json-schema.org/draft/2020-12/schema",
     "oneOf": [
@@ -190,7 +191,7 @@ TEST(Compiler_output_simple, success_oneof_1) {
   EXPECT_ANNOTATION_COUNT(output, 0);
 }
 
-TEST(Compiler_output_simple, fail_string) {
+TEST(Output_simple, fail_string) {
   const sourcemeta::core::JSON schema{sourcemeta::core::parse_json(R"JSON({
     "$schema": "https://json-schema.org/draft/2020-12/schema",
     "type": "string"
@@ -220,7 +221,7 @@ TEST(Compiler_output_simple, fail_string) {
   EXPECT_ANNOTATION_COUNT(output, 0);
 }
 
-TEST(Compiler_output_simple, fail_string_over_ref) {
+TEST(Output_simple, fail_string_over_ref) {
   const sourcemeta::core::JSON schema{sourcemeta::core::parse_json(R"JSON({
     "$schema": "https://json-schema.org/draft/2020-12/schema",
     "$ref": "#/$defs/string",
@@ -254,7 +255,7 @@ TEST(Compiler_output_simple, fail_string_over_ref) {
   EXPECT_ANNOTATION_COUNT(output, 0);
 }
 
-TEST(Compiler_output_simple, fail_string_with_matching_base) {
+TEST(Output_simple, fail_string_with_matching_base) {
   const sourcemeta::core::JSON schema{sourcemeta::core::parse_json(R"JSON({
     "$schema": "https://json-schema.org/draft/2020-12/schema",
     "$ref": "#/$defs/string",
@@ -290,7 +291,7 @@ TEST(Compiler_output_simple, fail_string_with_matching_base) {
   EXPECT_ANNOTATION_COUNT(output, 0);
 }
 
-TEST(Compiler_output_simple, fail_string_with_non_matching_base) {
+TEST(Output_simple, fail_string_with_non_matching_base) {
   const sourcemeta::core::JSON schema{sourcemeta::core::parse_json(R"JSON({
     "$schema": "https://json-schema.org/draft/2020-12/schema",
     "$ref": "#/$defs/string",
@@ -325,7 +326,7 @@ TEST(Compiler_output_simple, fail_string_with_non_matching_base) {
   EXPECT_ANNOTATION_COUNT(output, 0);
 }
 
-TEST(Compiler_output_simple, fail_oneof_1) {
+TEST(Output_simple, fail_oneof_1) {
   const sourcemeta::core::JSON schema{sourcemeta::core::parse_json(R"JSON({
     "$schema": "https://json-schema.org/draft/2020-12/schema",
     "oneOf": [
@@ -357,7 +358,7 @@ TEST(Compiler_output_simple, fail_oneof_1) {
   EXPECT_ANNOTATION_COUNT(output, 0);
 }
 
-TEST(Compiler_output_simple, fail_not_1) {
+TEST(Output_simple, fail_not_1) {
   const sourcemeta::core::JSON schema{sourcemeta::core::parse_json(R"JSON({
     "$schema": "https://json-schema.org/draft/2020-12/schema",
     "not": {
@@ -388,7 +389,7 @@ TEST(Compiler_output_simple, fail_not_1) {
   EXPECT_ANNOTATION_COUNT(output, 0);
 }
 
-TEST(Compiler_output_simple, fail_not_not_1) {
+TEST(Output_simple, fail_not_not_1) {
   const sourcemeta::core::JSON schema{sourcemeta::core::parse_json(R"JSON({
     "$schema": "https://json-schema.org/draft/2020-12/schema",
     "not": {
@@ -421,7 +422,7 @@ TEST(Compiler_output_simple, fail_not_not_1) {
   EXPECT_ANNOTATION_COUNT(output, 0);
 }
 
-TEST(Compiler_output_simple, fail_anyof_1) {
+TEST(Output_simple, fail_anyof_1) {
   const sourcemeta::core::JSON schema{sourcemeta::core::parse_json(R"JSON({
     "$schema": "https://json-schema.org/draft/2020-12/schema",
     "allOf": [
@@ -460,7 +461,7 @@ TEST(Compiler_output_simple, fail_anyof_1) {
   EXPECT_ANNOTATION_COUNT(output, 0);
 }
 
-TEST(Compiler_output_simple, fail_anyof_2) {
+TEST(Output_simple, fail_anyof_2) {
   const sourcemeta::core::JSON schema{sourcemeta::core::parse_json(R"JSON({
     "$schema": "https://json-schema.org/draft/2020-12/schema",
     "anyOf": [
@@ -494,7 +495,7 @@ TEST(Compiler_output_simple, fail_anyof_2) {
   EXPECT_ANNOTATION_COUNT(output, 0);
 }
 
-TEST(Compiler_output_simple, annotations_success_1) {
+TEST(Output_simple, annotations_success_1) {
   const sourcemeta::core::JSON schema{sourcemeta::core::parse_json(R"JSON({
     "$schema": "https://json-schema.org/draft/2020-12/schema",
     "title": "My Schema",
@@ -561,7 +562,7 @@ TEST(Compiler_output_simple, annotations_success_1) {
                           sourcemeta::core::JSON{true});
 }
 
-TEST(Compiler_output_simple, annotations_success_2) {
+TEST(Output_simple, annotations_success_2) {
   const sourcemeta::core::JSON schema{sourcemeta::core::parse_json(R"JSON({
     "$schema": "https://json-schema.org/draft/2020-12/schema",
     "anyOf": [
@@ -596,7 +597,7 @@ TEST(Compiler_output_simple, annotations_success_2) {
                           sourcemeta::core::JSON{"Third branch"});
 }
 
-TEST(Compiler_output_simple, annotations_success_3) {
+TEST(Output_simple, annotations_success_3) {
   const sourcemeta::core::JSON schema{sourcemeta::core::parse_json(R"JSON({
     "$schema": "https://json-schema.org/draft/2020-12/schema",
     "not": {
@@ -622,7 +623,7 @@ TEST(Compiler_output_simple, annotations_success_3) {
   EXPECT_ANNOTATION_COUNT(output, 0);
 }
 
-TEST(Compiler_output_simple, annotations_success_4) {
+TEST(Output_simple, annotations_success_4) {
   const sourcemeta::core::JSON schema{sourcemeta::core::parse_json(R"JSON({
     "$schema": "https://json-schema.org/draft/2020-12/schema",
     "contains": { "type": "boolean" }
@@ -652,7 +653,7 @@ TEST(Compiler_output_simple, annotations_success_4) {
                           sourcemeta::core::JSON{2});
 }
 
-TEST(Compiler_output_simple, annotations_success_5) {
+TEST(Output_simple, annotations_success_5) {
   const sourcemeta::core::JSON schema{sourcemeta::core::parse_json(R"JSON({
     "$schema": "https://json-schema.org/draft/2020-12/schema",
     "contains": { "type": "boolean" }
@@ -682,7 +683,7 @@ TEST(Compiler_output_simple, annotations_success_5) {
                           sourcemeta::core::JSON{2});
 }
 
-TEST(Compiler_output_simple, annotations_success_6) {
+TEST(Output_simple, annotations_success_6) {
   const sourcemeta::core::JSON schema{sourcemeta::core::parse_json(R"JSON({
     "$schema": "https://json-schema.org/draft/2020-12/schema",
     "contains": { "type": "boolean" }
@@ -712,7 +713,7 @@ TEST(Compiler_output_simple, annotations_success_6) {
                           sourcemeta::core::JSON{1});
 }
 
-TEST(Compiler_output_simple, annotations_success_7) {
+TEST(Output_simple, annotations_success_7) {
   const sourcemeta::core::JSON schema{sourcemeta::core::parse_json(R"JSON({
     "$schema": "https://json-schema.org/draft/2020-12/schema",
     "contains": { "type": "boolean" }
@@ -740,7 +741,7 @@ TEST(Compiler_output_simple, annotations_success_7) {
                           sourcemeta::core::JSON{0});
 }
 
-TEST(Compiler_output_simple, annotations_success_8) {
+TEST(Output_simple, annotations_success_8) {
   const sourcemeta::core::JSON schema{sourcemeta::core::parse_json(R"JSON({
     "$schema": "https://json-schema.org/draft/2020-12/schema",
     "unevaluatedItems": true
@@ -769,7 +770,7 @@ TEST(Compiler_output_simple, annotations_success_8) {
                           0, sourcemeta::core::JSON{true});
 }
 
-TEST(Compiler_output_simple, annotations_success_9) {
+TEST(Output_simple, annotations_success_9) {
   const sourcemeta::core::JSON schema{sourcemeta::core::parse_json(R"JSON({
     "$schema": "https://json-schema.org/draft/2020-12/schema",
     "contains": { "type": "boolean" },
@@ -805,7 +806,7 @@ TEST(Compiler_output_simple, annotations_success_9) {
                           0, sourcemeta::core::JSON{true});
 }
 
-TEST(Compiler_output_simple, annotations_success_10) {
+TEST(Output_simple, annotations_success_10) {
   const sourcemeta::core::JSON schema{sourcemeta::core::parse_json(R"JSON({
     "$schema": "https://json-schema.org/draft/2020-12/schema",
     "contains": { "type": "number", "title": "Test" }
@@ -835,7 +836,7 @@ TEST(Compiler_output_simple, annotations_success_10) {
                           0, sourcemeta::core::JSON{"Test"});
 }
 
-TEST(Compiler_output_simple, annotations_failure_1) {
+TEST(Output_simple, annotations_failure_1) {
   const sourcemeta::core::JSON schema{sourcemeta::core::parse_json(R"JSON({
     "$schema": "https://json-schema.org/draft/2020-12/schema",
     "title": "Foo",
@@ -858,7 +859,7 @@ TEST(Compiler_output_simple, annotations_failure_1) {
   EXPECT_ANNOTATION_COUNT(output, 0);
 }
 
-TEST(Compiler_output_simple, fail_stacktrace) {
+TEST(Output_simple, fail_stacktrace) {
   const sourcemeta::core::JSON schema{sourcemeta::core::parse_json(R"JSON({
     "$schema": "https://json-schema.org/draft/2019-09/schema",
     "properties": {
@@ -903,7 +904,7 @@ The object value was not expected to define unevaluated properties
 )JSON");
 }
 
-TEST(Compiler_output_simple, fail_stacktrace_with_indentation) {
+TEST(Output_simple, fail_stacktrace_with_indentation) {
   const sourcemeta::core::JSON schema{sourcemeta::core::parse_json(R"JSON({
     "$schema": "https://json-schema.org/draft/2019-09/schema",
     "properties": {
