+++ test/packaging/find_package/CMakeLists.txt
@@ -10,3 +10,4 @@ target_link_libraries(blaze_hello PRIVATE sourcemeta::core::jsonschema)
 target_link_libraries(blaze_hello PRIVATE sourcemeta::blaze::compiler)
 target_link_libraries(blaze_hello PRIVATE sourcemeta::blaze::evaluator)
 target_link_libraries(blaze_hello PRIVATE sourcemeta::blaze::linter)
+target_link_libraries(blaze_hello PRIVATE sourcemeta::blaze::output)
